Renames debbuild to rpmbuild
